tests: Move all code from exit.go to tests
Both the functions in exit.go are used only in tests.
Hence, move these to main_test.go and get rid of exit.go.

Fixes #402
